salam

+++ src/main/java/com/exemple/quiz_app/auth/repository/UserRepository.java
@@ -1,28 +1,17 @@
+// auth/repository/UserRepository.java
 package com.exemple.quiz_app.auth.repository;
-
-// auth/repository/UserRepository.java (version minimaliste)
 
 import com.exemple.quiz_app.auth.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, BigInteger> {
 
-    /**
-     * Trouver un utilisateur par son email
-     */
     Optional<User> findByEmail(String email);
-
-    /**
-     * Vérifier si un email existe déjà
-     */
     boolean existsByEmail(String email);
-
-    /**
-     * Vérifier si un utilisateur existe par ID
-     */
-    boolean existsById(Long id);
+    boolean existsById(BigInteger id);
 }
